+++ PatchMatch/PM.h
@@ -8,7 +8,7 @@ struct PMatch
 {
   Mat imgS,imgT;
   Mat vlS,vlT;
-  vector< vector<Point2d> > fS,initFs;
+  vector< vector<Point2i> > fS,initFs;
   vector< vector<int> >  fGood;
   bool loaded;
   int sw,sh,tw,th;
@@ -17,10 +17,10 @@ struct PMatch
   int sqrPsize,lhPatch,rhPatch;
   PMatch();
   void _vl();
-  bool _cplt(int x,int y,Point2d ofs);
+  bool _cplt(int x,int y,Point2i ofs);
   int _moveDown(int x,int y);
   int _moveRight(int x,int y);
-  int _cal(int x,int y,Point2d pnt,int pmax);
+  int _cal(int x,int y,Point2i pnt,int pmax);
   int load(Mat _s,Mat _t);
   int reload(Mat _s,bool _wh);
   int reset();
